requirements.txt
addict==2.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
einops==0.8.0
filelock==3.13.1
fonttools==4.53.1
fsspec==2024.2.0
huggingface-hub==0.23.4
idna==3.7
imageio==2.34.2
importlib_metadata==8.0.0
importlib_resources==6.4.0
Jinja2==3.1.3
kiwisolver==1.4.5
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
mmcv==2.2.0
mmengine==0.10.4
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
opencv-python==4.8.1.78
packaging==24.1
pillow==10.2.0
platformdirs==4.2.2
protobuf==5.27.2
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
safetensors==0.4.3
scikit-image==0.24.0
scipy==1.13.1
six==1.16.0
sympy==1.12
tensorboardX==2.6.2.2
termcolor==2.4.0
tifffile==2024.7.2
timm==1.0.7
tomli==2.0.1
torch==2.3.0+cu121
torchaudio==2.3.0+cu121
torchvision==0.18.0+cu121
tqdm==4.66.4
triton==2.3.0
typing_extensions==4.9.0
urllib3==2.2.2
yapf==0.40.2
zipp==3.19.2
